+++ src/com/itacademy/firstgit/MainLesson5_01.java
@@ -22,15 +22,15 @@ public class MainLesson5_01 {
 //		Animal[] animals = {murzik,bobik,funtik};
 
         Cat vaska = new Cat();
-        vaska.setAge(5);
+        vaska.setAge((Integer)5);
         vaska.setName("Vaska");
 
         Dog sharik = new Dog();
-        sharik.setAge(4);
+        sharik.setAge((Integer)4);
         sharik.setName("Sharik");
 
         Pig borka = new Pig();
-        borka.setAge(4);
+        borka.setAge((Integer)4);
         borka.setName("Borka");
 
         List<Animal> cats = new ArrayList<>();// смотрит что слево
